net: phy: ti: Allow the driver to be more configurable
Not all devices use the same internal delay or fifo depth.
Add the ability to set the internal delay for rx or tx and the
fifo depth via the devicetree.  If the value is not set in the
devicetree then set the delay to the default.

If devicetree is not used then use the default defines within the
driver.

Change-Id: I269095cf3491af63ba3c4185d02b85da894e41a6
Signed-off-by: Dan Murphy <dmurphy@ti.com>
Tested-by: Michal Simek <michal.simek@xilinx.com>
[mugunthanvnm@ti.com: Fix build error for non DM_ETH]
Signed-off-by: Mugunthan V N <mugunthanvnm@ti.com>
Signed-off-by: Ravi Babu <ravibabu@ti.com>