arm: dra7xx: Assign omap_vcores based on board type
Based on commits e7eb8f938734 and 4e295e047f4e branch ti-u-boot-2016.05

Currently omap_vcores which holds pmic data is being assigned based
on the SoC type. PMIC is not a part of SoC. It is logical to
to assign omap_vcores based on board type. Hence over ride the
vcores_update function and assign omap_vcores based on the board type.

Reported-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Keerthy <j-keerthy@ti.com>
Signed-off-by: Ravi Babu <ravibabu@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>