power: twl6030: Enable VAUX1 for eMMC power, depending on BOOT2 value
commit d7b6a75497ea2a33a1d2e4b8fa4d98dc8c7e876b branch ti-u-boot-2016.05

This enables the VAUX1 supply, used for eMMC power in standard configurations.
Its voltage is determined by the value of the BOOT2 pin of the TWL6030.

Note that the TWL6030 might already have enabled this regulator at startup
(depending on the value of the BOOT3 pin of the TWL6030), according to the
TWL6030 datasheet.

Signed-off-by: Paul Kocialkowski <contact@paulk.fr>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>