board: ti: dra7: route the MMC2 lines to the eMMC on DRA71 evm
On a DRA71 evm, the MMC2 lines are not directly connected to the eMMC. In
order to allow for the GPMC boot, it goes through an external MUX whose
selection line comes from the PCF8575 GPIO expander. To route the lines to
the eMMC, the select line must driven low.

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>