remoteproc: consolidate commons in a header file
There are many common definitions for IPU and DSP
remoteproc driver. To support early boot for a different
memory map, address and few other definitions will have to be
changed at multiple places.

All #defines are now in a single header file, so that changes
to be done in one place.

Signed-off-by: Ramprasad N <x0038811@ti.com>